backend/pyproject.toml
[project]
name = "plantpal-backend"
version = "0.2.0"
description = "PlantPal – Indoor Plant Care Tracker API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "pydantic-settings>=2.4.0",
    "redis>=5.0.0",
    "passlib>=1.7.4",
    "bcrypt<4.1,>=4.0.1",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.9",
    "tenacity>=9.0.0",
    "anyio>=4.4.0",
    "typer>=0.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
